+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next
 
 * Fix #6 - make tab say 'Destiny Raid Status' instead of 'DRS'.
+* Fix #7 - account for deleted characters.
 * Fix #8 - add link to Destiny Status and Destiny Tracker for each gamertag.
 * Semi-fix #3 - add some @media queries
 * Fixes to gamertag input (empty, whitespace).